requirements.txt
streamlit
sympy
numpy
pandas
google-genai
google-generativeai
python-dotenv